+++ CHANGELOG.md
@@ -17,6 +17,11 @@ All notable changes to this project will be documented in this file.
 - 'loadParameters' returns its success
 - 'sendParameters' can control if sent data should be saved directly by CSK_Module_PersistentData
 - Added UI icon and browser tab information
+- Info in UI if image type does not fit to canny or blur filter
+
+### Bugfix
+- Error if module is not active but 'getInstancesAmount' was called
+- processInstanceNUM did not work after deregistering from event to process images
 
 ## Release 1.0.0
 - Initial commit